Split absolute and relative path examples
The concern before this one already has split sections.

Rather than a block with two mixed concerns, and a need for "or"  text in a mixed-concern example attempting to demonstrate two things at once,
having two blocks makes it much more obvious, distinct, and reference-able (when the reader is only interested in one aspect).

It also simplifies the NU_LIB_DIRS concern a lot. Instead of a long block headline with a lot of mixed information, we do the same as the other concern earlier in the document:
note the information in an important note block.